# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MOVO API
  description: "API per il servizio di car sharing elettrico MOVO, operativo a Trento."
  version: 1.0.0
  contact:
    name: MOVO Tech
    email: api@movotn.it
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Sviluppo locale
  - url: https://api.movotn.it/v1
    description: Produzione
...
security:
  - BearerAuth: []
paths:
...
  /users/me:
    get:
      operationId: getProfile
      tags: [Profilo]
      summary: Profilo utente corrente
      responses:
        "200":
          description: Dati profilo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      operationId: updateProfile
      tags: [Profilo]
      summary: Aggiorna profilo
      description: Aggiorna solo i campi specificati. Email e codice fiscale non modificabili.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                language:
                  type: string
                  enum: [it, en, de]
      responses:
        "200":
          description: Profilo aggiornato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/documents:
    post:
      operationId: uploadDocuments
      tags: [Profilo]
      summary: Carica/aggiorna documenti
      description: |
        Per utenti passeggeri che vogliono abilitarsi alla guida,
        o per aggiornare documenti in scadenza.

        Dopo l'upload, lo status passa a `pending_approval`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                driving_license:
                  type: string
                  format: binary
                identity_document:
                  type: string
                  format: binary
      responses:
        "200":
          description: Documenti caricati
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending_approval]
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /rentals/unlock:
    post:
      operationId: unlockVehicle
      tags: [Noleggi]
      summary: Sblocca veicolo e inizia noleggio
      description: |
        Sblocca il veicolo prenotato e avvia il noleggio.
        Richiede prossimità al veicolo (~50m).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [booking_id]
              properties:
                booking_id:
                  type: string
                  format: uuid
                user_location:
                  $ref: "#/components/schemas/GeoPoint"
      responses:
        "200":
          description: Veicolo sbloccato, noleggio attivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Utente troppo lontano dal veicolo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          $ref: "#/components/responses/NotFound"
  /rentals/active:
    get:
      operationId: getActiveRental
      tags: [Noleggi]
      summary: Noleggio attivo
      responses:
        "200":
          description: Noleggio in corso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Nessun noleggio attivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /rentals/{rentalId}/end:
    post:
      operationId: endRental
      tags: [Noleggi]
      summary: Termina noleggio
      description: |
        Chiude il noleggio. Il veicolo deve essere in un'area di sosta autorizzata.
        Il costo finale viene addebitato sul metodo di pagamento predefinito.
      parameters:
        - $ref: "#/components/parameters/rentalId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                end_location:
                  $ref: "#/components/schemas/GeoPoint"
                photos:
                  type: array
                  maxItems: 4
                  items:
                    type: string
                    format: byte
      responses:
        "200":
          description: Noleggio concluso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Veicolo non in area di sosta valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          $ref: "#/components/responses/NotFound"
  /rentals/history:
    get:
      operationId: getRentalHistory
      tags: [Noleggi]
      summary: Storico noleggi
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Lista noleggi
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RentalSummary"
                  pagination:
                    $ref: "#/components/schemas/CursorPagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  parameters:
...
    rentalId:
      name: rentalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT ottenuto da /auth/login o /auth/login/google

    OAuth2Google:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            email: Accesso email utente
            profile: Accesso dati profilo base
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, first_name, last_name, role, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        fiscal_code:
          type: string
        phone:
          type: string
        address:
          type: string
        language:
          type: string
          enum: [it, en, de]
          default: it
        role:
          type: string
          enum: [user, operator, admin]
        status:
          type: string
          enum: [active, pending_approval, suspended, blocked]
        driving_enabled:
          type: boolean
        driving_license_verified:
          type: boolean
        driving_license_expiry:
          type: string
          format: date
          nullable: true
        identity_document_expiry:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    GeoPoint:
      type: object
      required: [type, coordinates]
      description: GeoJSON Point (RFC 7946)
      properties:
        type:
          type: string
          enum: [Point]
        coordinates:
          type: array
          description: "[longitude, latitude]"
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: double
          example: [11.1257, 46.0664]
...
    VehicleMapItem:
      type: object
      required: [id, model, license_plate, location, battery_level, status]
      properties:
        id:
          type: string
          format: uuid
        model:
          type: string
          example: "Fiat 500e"
        license_plate:
          type: string
          example: "TN123AB"
        location:
          $ref: "#/components/schemas/GeoPoint"
        battery_level:
          type: integer
          minimum: 0
          maximum: 100
        status:
          type: string
          enum: [available, reserved, in_use]
        base_price_per_minute:
          type: number
          description: Centesimi/minuto
          example: 35
...
    Rental:
      type: object
      required: [id, user_id, vehicle_id, status, started_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        vehicle_id:
          type: string
          format: uuid
        vehicle:
          $ref: "#/components/schemas/VehicleMapItem"
        status:
          type: string
          enum: [active, paused, completed]
        started_at:
          type: string
          format: date-time
        start_location:
          $ref: "#/components/schemas/GeoPoint"
        current_cost_cents:
          type: integer
        paused_at:
          type: string
          format: date-time
          nullable: true
    RentalSummary:
      allOf:
        - $ref: "#/components/schemas/Rental"
        - type: object
          properties:
            ended_at:
              type: string
              format: date-time
            end_location:
              $ref: "#/components/schemas/GeoPoint"
            duration_minutes:
              type: integer
            distance_km:
              type: number
            total_cost_cents:
              type: integer
            discount_applied_cents:
              type: integer
            final_cost_cents:
              type: integer

    # ── Payments --------------------------------------------------------──────
...
    CursorPagination:
      type: object
      required: [has_more, limit]
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        limit:
          type: integer
    ProblemDetails:
      type: object
      required: [type, title, status]
      description: Formato errore RFC 7807
      properties:
        type:
          type: string
          format: uri
          example: "https://api.movotn.it/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 400
        detail:
          type: string
        instance:
          type: string
          format: uri
  responses:
    BadRequest:
      description: Richiesta non valida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Autenticazione richiesta o token non valido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    NotFound:
      description: Risorsa non trovata
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...